+++ commands/help.js
@@ -29,18 +29,22 @@ module.exports = {
 				command = client.commands.get(command);
 
 				var aliases = '';
-				var syntax = '``' + command.syntax + '``';
+				var syntax = '```' + command.syntax + '```';
+				if (command.flags) syntax += '```' + command.flags + '```';
+
 				if (command.aliases) {
 						for (var alias in command.aliases) {
 						aliases += '``' + prefix + command.aliases[alias] + '`` ';
 					}
 				} else aliases = 'None'
 
+				var example = '```' + command.example + '```' || 'None';
+
 				var embed = new Discord.MessageEmbed()
 					.setTitle(`Help: ${prefix}${command.name}`)
 					.setThumbnail(settings.BOT_IMG)
 					.setColor(colors.BLUE)
-					.setDescription(`**Description:**\n${command.description || 'No description'}\n**Can be used by:**\n${command.uses || 'Not set'}\n**Syntax:**\n${syntax}\n**Aliases:**\n${aliases}`)
+					.setDescription(`**Description:**\n${command.description || 'No description'}\n**Can be used by:**\n${command.uses || 'Not set'}\n**Syntax:**\n${syntax}\n**Aliases:**\n${aliases}\n**Example:**\n${example}`)
 					.setFooter('CompliBot', settings.BOT_IMG);
 
 			} else return warnUser(message, 'Please provide valid commands & do not use aliases');