+++ commands/textureTools/palette.js
@@ -85,7 +85,7 @@ module.exports = {
 			}
 
 			if (found) await palette(message, url);
-			else return
+			else return warnUser(message,'No image found in the 10 previous messages.');
 		}
 	}
 }