+++ commands/textureTools/tile.js
@@ -77,7 +77,7 @@ module.exports = {
 			}
 
 			if (found) await tile(message, url, type);
-			else return
+			else return warnUser(message,'No image found in the 10 previous messages.');
 		}
 	}
 }